+++ lib/aws-activejob-sqs.rb
@@ -1,5 +1,6 @@
 # frozen_string_literal: true
 
+require 'active_job'
 require_relative 'active_job/queue_adapters/sqs_adapter'
 require_relative 'active_job/queue_adapters/sqs_adapter/params'
 require_relative 'active_job/queue_adapters/sqs_async_adapter'
@@ -11,6 +12,7 @@ require_relative 'aws/active_job/sqs/lambda_handler'
 
 module Aws
   module ActiveJob
+    # ActiveJob Adapter and backend queueing using AWS SQS.
     module SQS
       VERSION = File.read(File.expand_path('../VERSION', __dir__)).strip
 
@@ -19,7 +21,7 @@ module Aws
         @config ||= Configuration.new
       end
 
-      # @yield Configuration
+      # @yield [Configuration] the (singleton) Configuration
       def self.configure
         yield(config)
       end