+++ lib/aws/active_job/sqs/executor.rb
@@ -31,21 +31,27 @@ module Aws
 
         def initialize(options = {})
           @executor = Concurrent::ThreadPoolExecutor.new(DEFAULTS.merge(options))
-          @retry_standard_errors = options[:retry_standard_errors]
           @logger = options[:logger] || ActiveSupport::Logger.new($stdout)
           @task_complete = Concurrent::Event.new
+          @post_mutex = Mutex.new
+
+          @error_handler = options[:error_handler]
+          @error_queue = Thread::Queue.new
+          @error_handler_thread = Thread.new(&method(:handle_errors))
+          @error_handler_thread.abort_on_exception = true
+          @error_handler_thread.report_on_exception = false
+          @shutting_down = Concurrent::AtomicBoolean.new(false)
         end
 
         def execute(message)
-          post_task(message)
-        rescue Concurrent::RejectedExecutionError
-          # no capacity, wait for a task to complete
-          @task_complete.reset
-          @task_complete.wait
-          retry
+          @post_mutex.synchronize do
+            _execute(message)
+          end
         end
 
         def shutdown(timeout = nil)
+          @shutting_down.make_true
+
           run_hooks_for(:stop)
           @executor.shutdown
           clean_shutdown = @executor.wait_for_termination(timeout)
@@ -57,40 +63,70 @@ module Aws
                          'the queue and can be ru-run once their visibility timeout ' \
                          'passes.'
           end
+          @error_queue.push(nil) # process any remaining errors and then terminate
+          @error_handler_thread.join unless @error_handler_thread == Thread.current
+          @shutting_down.make_false
         end
 
         private
 
-        def post_task(message)
-          @executor.post(message) do |message|
-            job = JobRunner.new(message)
-            @logger.info("Running job: #{job.id}[#{job.class_name}]")
-            job.run
-            message.delete
-          rescue Aws::Json::ParseError => e
-            @logger.error "Unable to parse message body: #{message.data.body}. Error: #{e}."
-          rescue StandardError => e
-            job_msg = job ? "#{job.id}[#{job.class_name}]" : 'unknown job'
-            @logger.info "Error processing job #{job_msg}: #{e}"
-            @logger.debug e.backtrace.join("\n")
+        def _execute(message)
+          post_task(message)
+        rescue Concurrent::RejectedExecutionError
+          # no capacity, wait for a task to complete
+          @task_complete.reset
+          @task_complete.wait
+          retry
+        end
 
-            if @retry_standard_errors && !job.exception_executions?
-              @logger.info(
-                'retry_standard_errors is enabled and job has not ' \
-                "been retried by Rails.  Leaving #{job_msg} in the queue."
-              )
-            else
-              message.delete
-            end
-          ensure
-            @task_complete.set
+        def post_task(message)
+          @executor.post(message) do |msg|
+            execute_task(msg)
           end
+        end
+
+        def execute_task(message)
+          job = JobRunner.new(message)
+          @logger.info("Running job: #{job.id}[#{job.class_name}]")
+          job.run
+          message.delete
+        rescue JSON::ParserError => e
+          @logger.error "Unable to parse message body: #{message.data.body}. Error: #{e}."
+        rescue StandardError => e
+          handle_standard_error(e, job, message)
+        ensure
+          @task_complete.set
+        end
+
+        def handle_standard_error(error, job, message)
+          job_msg = job ? "#{job.id}[#{job.class_name}]" : 'unknown job'
+          @logger.info "Error processing job #{job_msg}: #{error}"
+          @logger.debug error.backtrace.join("\n")
+
+          @error_queue.push([error, message])
         end
 
         def run_hooks_for(event_name)
           return unless (hooks = self.class.lifecycle_hooks[event_name])
 
           hooks.each(&:call)
+        end
+
+        # run in the @error_handler_thread
+        def handle_errors
+          # wait until errors are placed in the error queue
+          while ((exception, message) = @error_queue.pop)
+            raise exception unless @error_handler
+
+            @error_handler.call(exception, message)
+
+          end
+        rescue StandardError => e
+          @logger.info("Unhandled exception executing jobs in poller: #{e}.")
+          @logger.info('Shutting down executor')
+          shutdown unless @shutting_down.true?
+
+          raise e # re-raise the error, terminating the application
         end
       end
     end