+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 Unreleased Changes
 ------------------
 
+* Feature - Support polling on multiple queues. (#4)
+* Feature - Support running without Rails. (#5)
+* Feature - Replace `retry_standard_errors` with `poller_error_handler`. (#6)
+* Feature - Support per queue configuration. (#4)
+* Feature - Support loading global and queue specific configuration from ENV. (#3)
+
 0.1.1 (2024-12-02)
 ------------------